feat: OTLP write
Implements a Canary OTLP write and a simple loop. Much of this is
placeholder and will need to be re-factored, also fails linter. Intended
to get a working OTLP metric write setup in the docker compose
environment to enable further development

Logic was to implement a seperate meter provider in the same namespace
as the main o11y-canary provider