feat: TLS (#17)
* feat: TLS

adds TLS support, making it specific to each URL for ingest/query. also
allows a global ingest OR query TLS setting that URLs will inherit if
they dont have specifically defined

as a result had to refactor a lot of the URL logic in main to accomodate
this!

* docs: add example config

* docs: point to test repo

* ci: add local certs

and try to fix canarying problems

* fix: use certs for otel col

* fix: force flush OTLP metrics

force flushing solves a lot of export issues

* fix: properly register metrics

and improve returns of c.Query()

* test: make tests more common sense?